requirements.txt
# Web Framework
flask>=3.0.0

# LangChain packages
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0

# Vector Database
faiss-cpu>=1.7.4

# PDF Processing
pypdfium2>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Numerical Computing
numpy>=1.24.0

